requirements.txt
nltk==3.4.1
numpy==1.16.2
pandas==0.24.2
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
tabulate==0.8.3
tqdm==4.31.1
xmltodict==0.12.0
